A complete rework of the skeleton DSL
This was done in order to provide a better look and feel to the skeleton
form. As time went on at KISSmetrics, I realized we had deeper needs
than what was initially required.

An Options serializer will be implemented soon to allow for options
requests to spit out only the required information for a specific end
point. It will be Swagger-esque but a subset of what is defined. Only
the required models will be included along with omission of the
information

As such the this is a work in progress.